Added notebook sorter (#184)
Added the preference to sort notebooks by:
- Alphabetical name
- Creation date (essentially, notebook ID which increments by time)

Both the "Manage Notebook" and the navigation drawer's notebook list are affected by the notebook sorting. Please note that sorting ofnotebook list in the nav drawer only gets updated when the app restarts. This is because nav. drawer is, so far, designed to re-render only during launch.

The sorter button in the "Manage Notebook" menu only updates when the ViewModel gets updated (e.g., when the user clicks on the hamburger menu twice). I tried using forced re-rendering of UI, but it was fruitless.

At least, for now, we can sort notebooks alphabetically via the settings.

This commit might resolve https://github.com/quillpad/quillpad/issues/184.