Use the note title as a fallback for the reminder title (#424)
Currently, if there is no reminder title, the notification won't show any thing to help locate what it is about (can't open the note on my devices), and the bell icon and scheduled date time in the note are also cleaned up.

I'd like to suggest use the note title instead, if no reminder title. It is a easy fallback, and practical.

    Normally, the reminder is most likely with the same title. Repeat inputting isn't necessary.
    There is a default note title. People will change it to be meaningful if the note is important.

It's a kind of workaround for #87.